Game Manager v 1.0
Game manager purpose:
1. Tracking and storing the game state
a. Functions for changing the game state (StartGame() , PauseGame(), GameOver(), OnWinGame(), LoadCutscene(), LoadSomething())
b. Changing the game state,

2. Creating player/ players objects
3. Camera Setup
4. Communicating between manager scripts like: session , user, score etc etc

2.2 Building the Core Framework Scripts 13
There may also be several custom states that are applicable to particular types of
games, such as a state for showing a cutscene or perhaps a state for playing a special audio
clip.